+++ src/construo/SphereCoverage.cpp
@@ -102,9 +102,9 @@ contains(const Point& v) const
 
     //check the point for region changes against all polygon edges:
     int numRegionChanges = 0;
-    uint numVertices = static_cast<uint>(vertices.size());
+    size_t numVertices = static_cast<size_t>(vertices.size());
     const Point* prev = &vertices[numVertices-1];
-    for (uint i=0; i<numVertices; ++i)
+    for (size_t i=0; i<numVertices; ++i)
     {
         const Point* curr = &vertices[i];
         if ( ((*prev)[1] <= v[1]) && ((*curr)[1] > v[1]) )
@@ -129,7 +129,7 @@ contains(const Point& v) const
     return (numRegionChanges&0x1) != 0x0;
 }
 
-uint SphereCoverage::
+size_t SphereCoverage::
 overlaps(const Box& bBox) const
 {
     //check box against the bounding box first
@@ -139,9 +139,9 @@ overlaps(const Box& bBox) const
     /* check if any polygon edges intersect the box, and check the box's min
        vertex for region changes against all polygon edges: */
     int numRegionChanges = 0;
-    uint numVertices = static_cast<uint>(vertices.size());
+    size_t numVertices = static_cast<size_t>(vertices.size());
     const Point* prev = &vertices[numVertices-1];
-    for (uint vertexIndex=0; vertexIndex<numVertices; ++vertexIndex)
+    for (size_t vertexIndex=0; vertexIndex<numVertices; ++vertexIndex)
     {
         const Point* curr = &vertices[vertexIndex];
 
@@ -239,11 +239,11 @@ overlaps(const Box& bBox) const
     return (numRegionChanges&0x1)!=0x0 ? CONTAINS : SEPARATE;
 }
 
-uint SphereCoverage::
+size_t SphereCoverage::
 overlaps(const SphereCoverage& coverage) const
 {
     //must check overlap in the 4 spaces produced by periodicity
-    uint res = checkOverlap(coverage);
+    size_t res = checkOverlap(coverage);
     if (res != SEPARATE)
         return res;
 
@@ -305,7 +305,7 @@ fprintf(stderr, "<%f, %f>\n", vertex[0], vertex[1]);
     box.addPoint(vertex);
 }
 
-uint SphereCoverage::
+size_t SphereCoverage::
 checkOverlap(const SphereCoverage& coverage) const
 {
     //check box against the bounding box first
@@ -357,51 +357,51 @@ checkOverlap(const SphereCoverage& coverage) const
 }
 
 StaticSphereCoverage::
-StaticSphereCoverage(uint subdivisions, const Scope& scope)
+StaticSphereCoverage(size_t subdivisions, const Scope& scope)
 {
-    uint numSamples = pow(2, subdivisions) + 1;
-    uint lastSample = numSamples - 1;
+    size_t numSamples = pow(2, subdivisions) + 1;
+    size_t lastSample = numSamples - 1;
     Scope::Vertex* samples = new Scope::Vertex[numSamples];
-    uint numVertices = 4 * numSamples - 4;
+    size_t numVertices = 4 * numSamples - 4;
     vertices.reserve(numVertices);
 
     /* go through all the side-segments of the scope in order and produce
        samples on each segment through 'subdivisions' number of subdivisions.
        Each sample corresponsing to a vertex of the coverage */
-    static const uint remap[4] = {0, 1, 3, 2};
+    static const size_t remap[4] = {0, 1, 3, 2};
     Scope::Scalar radius       = scope.getRadius();
-    for (uint start=0; start<4; ++start)
+    for (size_t start=0; start<4; ++start)
     {
-        uint end            = (start+1) % 4;
+        size_t end            = (start+1) % 4;
         samples[0]          = scope.corners[remap[start]];
         samples[lastSample] = scope.corners[remap[end]];
 
         //iterate over block sizes to generate each subdivision level
-        for (uint blockSize=lastSample; blockSize>1; blockSize>>=1)
+        for (size_t blockSize=lastSample; blockSize>1; blockSize>>=1)
         {
-            uint halfBlockSize = blockSize>>1;
+            size_t halfBlockSize = blockSize>>1;
             //iterate over the blocks of a level to generate the mid points
-            for (uint block=0; block<lastSample; block+=blockSize)
+            for (size_t block=0; block<lastSample; block+=blockSize)
             {
                 Scope::Vertex& lower = samples[block];
                 Scope::Vertex& upper = samples[block+blockSize];
                 Scope::Vertex& mid   = samples[block+halfBlockSize];
                 //accumulate the norm while computing the mid-point
                 double  norm  = 0;
-                for (uint i=0; i<3; ++i)
+                for (size_t i=0; i<3; ++i)
                 {
                     mid[i] = (lower[i] + upper[i]) * 0.5;
                     norm += mid[i]*mid[i];
                 }
                 //normalize
                 norm = radius / sqrt(norm);
-                for (uint i=0; i<3; ++i)
+                for (size_t i=0; i<3; ++i)
                     mid[i] *= norm;
             }
         }
 
         //append the new samples as spherical vertices to the set
-        for (uint i=0; i<numSamples-1; ++i)
+        for (size_t i=0; i<numSamples-1; ++i)
         {
             Point newVertex = Converter::cartesianToSpherical(samples[i]);
             addVertex(newVertex);
@@ -425,29 +425,29 @@ fprintf(stderr, "[[%f, %f]]\n", shiftVec[0], shiftVec[1]);
 }
 
 StaticSphereCoverage::
-StaticSphereCoverage(uint subdivisions, const ImageCoverage* coverage,
+StaticSphereCoverage(size_t subdivisions, const ImageCoverage* coverage,
                      const ImageTransform* transform)
 {
     assert(coverage!=NULL && transform!=NULL);
 
-    uint numSamples     = pow(2, subdivisions) + 1;
+    size_t numSamples     = pow(2, subdivisions) + 1;
     double sampleStep   = 1.0/(numSamples-1);
-    uint numImgVertices = coverage->getNumVertices();
-    uint numVertices    = numImgVertices*numSamples - numImgVertices;
+    size_t numImgVertices = coverage->getNumVertices();
+    size_t numVertices    = numImgVertices*numSamples - numImgVertices;
     vertices.reserve(numVertices);
 
     /* go through all the side-segments of the image in order and produce
      samples on each segment through 'subdivisions' number of subdivisions.
      Each sample corresponsing to a vertex of the coverage */
-    for (uint start=0; start<numImgVertices; ++start)
+    for (size_t start=0; start<numImgVertices; ++start)
     {
-        uint end           = (start+1) % numImgVertices;
+        size_t end           = (start+1) % numImgVertices;
         const Point& lower = coverage->getVertex(start);
         const Point& upper = coverage->getVertex(end);
 
         //generate all the samples along that segment
         double alpha = 0.0;
-        for (uint i=0; i<numSamples-1; ++i, alpha+=sampleStep)
+        for (size_t i=0; i<numSamples-1; ++i, alpha+=sampleStep)
         {
             Point imgPoint((1.0-alpha)*lower[0] + alpha*upper[0],
                            (1.0-alpha)*lower[1] + alpha*upper[1]);
